refactor: ProductSummary API로직 api파일 안에서 관리하게 분리
