+++ src/pages/OrderPage/ProductSummary/ProductSummary.tsx
@@ -1,23 +1,11 @@
 import { useParams } from 'react-router-dom';
 import { useReactQueryFetch } from '@/hooks/useReactQueryFetch';
+import { fetchProductSummary } from '@/api/products';
 import * as S from './ProductSummary.styles';
-
-interface ProductSummaryData {
-  id: number;
-  name: string;
-  price: number;
-  imageURL: string;
-  brandName?: string;
-}
 
 const ProductSummary = () => {
   const { id } = useParams<{ id: string }>();
   const productId = parseInt(id ?? '', 10);
-
-  const url = isNaN(productId) ? null : `http://localhost:3000/api/products/${productId}/summary`;
-  if (!url) return null;
-  const { data, isLoading, error } = useReactQueryFetch<ProductSummaryData>(url);
-  const product = data?.data;
 
   if (isNaN(productId)) {
     return (
@@ -26,6 +14,12 @@ const ProductSummary = () => {
       </S.Wrapper>
     );
   }
+
+  const { data, isLoading, error } = useReactQueryFetch(['productSummary', productId], () =>
+    fetchProductSummary(productId)
+  );
+
+  const product = data?.data;
 
   if (isLoading) {
     return (
